--- a/src/components/transactions/history/details/getListEntries.tsx
+++ b/src/components/transactions/history/details/getListEntries.tsx
@@ -4,7 +4,7 @@ import { ReactNode } from 'react';
 import { formatNumber, FormatPreset } from '@app/utils';
 import { StatusListEntry } from '@app/components/transactions/components/StatusList/StatusList.tsx';
 import { getExplorerUrl, Network } from '@app/utils/network.ts';
-import { CombinedBridgeWalletTransaction, useMiningMetricsStore, useMiningStore } from '@app/store';
+import { CombinedBridgeWalletTransaction, useNodeStore, useMiningStore } from '@app/store';
 import { getTxStatusTitleKey, getTxTitle } from '@app/utils/getTxStatus.ts';
 import { EmojiAddressWrapper } from '@app/components/transactions/history/details/styles.ts';
 
@@ -42,7 +42,7 @@ function getPaymentReferenceValue(transaction: CombinedBridgeWalletTransaction):
     if (transaction.walletTransactionDetails.paymentReference) {
         return transaction.walletTransactionDetails.paymentReference;
     }
-    const currentBlockHeight = useMiningMetricsStore.getState().base_node_status.block_height;
+    const currentBlockHeight = useNodeStore.getState().base_node_status.block_height;
     if (!transaction.mined_in_block_height || !currentBlockHeight) {
         return i18n.t('common:pending');
     }