--- a/src/containers/floating/Settings/sections/connections/ConnectionStatus.tsx
+++ b/src/containers/floating/Settings/sections/connections/ConnectionStatus.tsx
@@ -4,11 +4,11 @@ import { IoCheckmarkCircle, IoCloudOfflineOutline } from 'react-icons/io5';
 import { ConnectionIcon } from '../../components/Settings.styles.tsx';
 import { Typography } from '@app/components/elements/Typography.tsx';
 import { useTranslation } from 'react-i18next';
-import { useMiningMetricsStore } from '@app/store/useMiningMetricsStore.ts';
+import { useNodeStore } from '@app/store/useNodeStore.ts';
 
 export default function ConnectionStatus() {
     const { t } = useTranslation('settings');
-    const isConnectedToTariNetwork = useMiningMetricsStore((s) => s.isNodeConnected);
+    const isConnectedToTariNetwork = useNodeStore((s) => s.isNodeConnected);
 
     return (
         <Stack direction="row" justifyContent="right" alignItems="center">