--- a/src/containers/navigation/components/VersionChip/ConnectedPulse/ConnectedPulse.tsx
+++ b/src/containers/navigation/components/VersionChip/ConnectedPulse/ConnectedPulse.tsx
@@ -1,8 +1,8 @@
 import { Dot, Pulse, Wrapper } from './styles';
-import { useMiningMetricsStore } from '@app/store';
+import { useNodeStore } from '@app/store';
 
 export default function ConnectedPulse({ size = 5 }: { size?: number }) {
-    const isConnectedToTariNetwork = useMiningMetricsStore((s) => s.isNodeConnected);
+    const isConnectedToTariNetwork = useNodeStore((s) => s.isNodeConnected);
     return (
         <Wrapper $size={size}>
             <Dot $isConnected={isConnectedToTariNetwork} $size={size} />