--- a/src/store/index.ts
+++ b/src/store/index.ts
@@ -14,3 +14,4 @@ export * from './useShellOfSecretsStore.ts';
 export * from './useSecurityStore.ts';
 export * from './useUIStore.ts';
 export * from './useWalletStore.ts';
+export * from './useNodeStore.ts';