Description
---
#### _get_network_state_
**(+ 80% calls reduction)**

* Remote node's network status fetched once per 45secs instead of once per 10secs
* Orphan check uses global network status instead of fetching it's own once per 30sec(btw increased this to 60secs)

#### _get_blocks_
**(~66% calls reduction)**
- Orphan check changed to be performed once per 60secs instead of once per 30secs
- telemetry service uses already defined flag

#### _list_connected_peers_
**(~ 99% calls reduction)**
- `list_connected_peers` called on demand, not once per 10 sec

Motivation and Context
---
* Remote nodes are overloaded with grpc calls that could be optimize

Frontend updates
---
Move node related properties from `miningMetricsStore` to `nodeStore`